Fix for non-deterministic failing tests (#669)
* Moved tests for `NoEntityFrameworkTests` into a different database, because two variants of `TodoItem` exist (where one is a subset of the other), causing tests to fail/succeed depending on their execution order.

A separate database is needed instead of schema, because `context.Database.EnsureCreated()` only checks for database existence and does not create additional tables when the database already exists.

Furthermore, I decided to remove fields from `NoEntityFrameworkExample.Models.TodoItem` that do not exist in the INSERT SQL statement in `NoEntityFrameworkExample.Services.TodoItemService` and because non-nullable CreatedDate causes insert to fail.

* Fixed another bug, where tests would sometimes fail

This occured when working in the solution for a while, building and rebuilding one or more projects individually. What happened is that the appsettings.json from NoEntityFrameworkCoreExample project would overwrite the one from JsonApiDotNetCoreExample in the output folder of JsonApiDotNetCoreExample. Causing that project to reference the wrong database.

I solved this by removing the unneeded dependency between these projects and by relying on automatic copying of appsettings by the project system. I removed the copied appsettings.json files from the test projects, which end up in the output directory anyway (because they reference the *Example projects).

* Fixed invalid logging configuration