Updated build scripts to use auto-discovery (#683)
* Fixed another cause for random test failures

The problem: NoEntityFrameworkTests project has a reference to UnitTests project, which has a reference to JsonApiDotNetCoreExampleTests project, which has a reference to JsonApiDotNetCoreExample project. The last one contains an appsettings.json file with connection string to JsonApiDotNetCoreExample database. So when building, this file may end up in the output directory of NoEntityFrameworkTests project, overwriting the appsettings.json (containing the connection string to JsonApiDotNetCoreNoEFCoreExample database) from NoEntityFrameworkExample project.

Next, if the NoEntityFrameworkTests testset runs first, it populates the JsonApiDotNetCoreExample database from models intended for the JsonApiDotNetCoreNoEFCoreExample database. And this causes subsequent test projects to fail.

* Updated build scripts to use auto-discovery

* Tried another workaround